requirements.txt
streamlit
scikit-learn
pandas
numpy<2.0
nltk
spacy
pdfplumber
python-docx
torch
torchvision
matplotlib
seaborn
plotly
tqdm
